+++ CHANGELOG.md
@@ -32,9 +32,10 @@ only at the last position
 - Add `else` clause to for loops
 - Filters are now evaluated when checking if/elif conditions
 - Allow `{{-` and `-}}` for whitespace management
-- Adds a `xml_escape` filter
+- Add `xml_escape` filter
 - Grave accent is no longer escaped in HTML, not really needed anymore
 - Add a `builtins` default feature that gate all filters/functions requiring additional dependencies
+- Add `unique` and `map` filter
 
 ## 0.11.20 (2018-11-14)