+++ CHANGELOG.md
@@ -37,6 +37,7 @@ only at the last position
 - Grave accent is no longer escaped in HTML, not really needed anymore
 - Add a `builtins` default feature that gate all filters/functions requiring additional dependencies
 - Add `unique` and `map` filter
+- Add a `timezone` attribute to the `date` filter
 
 ## 0.11.20 (2018-11-14)