+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 1.0.0 (unreleased)
+## 1.0.0 (2019-12-05)
 
 ### Breaking changes